Update xiris_melee.inc
I had an issue where self buffs described in the bot_clr_name.INI,  as downflags, were unable to recognize a "spell" type downflag if it was preceded by "item" or "alt"type downflags in the INI file order.  This was because the loop iterating the downflags variable assignments was just missing one last else statement to reassign the var dsType back to SPELL if indicated.   I don't know if I explained that using the right terms.  I didn't want to bother you on discord as you were DND but I knew you would see this here.